requirements.txt
Click==7.0
CoffeeScript==1.0.8
Flask==0.10.1
Flask-Assets==2.0
Flask-Markdown==0.3
flask-paginate==0.5.5
Frozen-Flask==0.15
itsdangerous==0.23
Jinja2==2.7.2
Markdown==3.1.1
MarkupSafe==1.1.1
PyExecJS==1.0.4
Pygments==2.5.2
pyScss==1.3.7
PyYAML==5.2
six==1.13.0
Unidecode==1.1.1
webassets==2.0
Werkzeug==0.9.4
feedgen==0.9.0
pytz==2019.3
